Add DataFormatFlags and DataFormatRange (plus docs)
Adds support for configuring the accelerometer data format, including
optional support for reading "justified" (i.e. unsigned) accelerometer
data as a U16x3 in addition to reading signed data as an I16x3.

Also allows configuring the accelerometer sensitivity, which is useful
for getting a less noisy signal for getting a gravity reading.

Additionally better documents all of the registers and aligns their
names with what's documented in the data sheet.