Add `invert_z_axis()` method
The `accelerometer` v0.12 release inverted (actually corrected) how the
orientation tracker handles the `z`-component reading when calculating
an orientation.

It was wrong because it was based off the Trellis M4, which mounts the
ADXL343 on the underside of the board.

This commit adds a convenient `invert_z_axis()` method to handle this
scenario.